requirements.txt
dataclasses
numpy
matplotlib
pandas
scipy
pillow
tabulate
prettyprinter
IPython
